refactor(permission): split rules/mod.rs into submodules
Split the 579-line monolithic rules/mod.rs into focused submodules:
- mod.rs (12 lines): module declarations + pub use re-exports
- builder.rs (129 lines): RuleBuilder + RuleBuilderError
- ruleset_builder.rs (112 lines): RuleSetBuilder + RuleSetBuilderError
- validation.rs (83 lines): validation functions + error types
- tests.rs (238 lines): #[cfg(test)] mod tests

All public APIs remain unchanged. No logic modifications.

Source: issue #219
Type: refactor
